Enum vcf2parquet_lib::name2data::ColumnData

pub enum ColumnData {
    Bool(MutableBooleanArray),
    Int(MutablePrimitiveArray<i32>),
    Float(MutablePrimitiveArray<f32>),
    String(MutableUtf8Array<i32>),
    ListBool(MutableListArray<i32, MutableBooleanArray>),
    ListInt(MutableListArray<i32, MutablePrimitiveArray<i32>>),
    ListFloat(MutableListArray<i32, MutablePrimitiveArray<f32>>),
    ListString(MutableListArray<i32, MutableUtf8Array<i32>>),
}

Variants

Bool(MutableBooleanArray)

Int(MutablePrimitiveArray<i32>)

Float(MutablePrimitiveArray<f32>)

String(MutableUtf8Array<i32>)

ListBool(MutableListArray<i32, MutableBooleanArray>)

ListInt(MutableListArray<i32, MutablePrimitiveArray<i32>>)

ListFloat(MutableListArray<i32, MutablePrimitiveArray<f32>>)

ListString(MutableListArray<i32, MutableUtf8Array<i32>>)

Implementations

impl ColumnData

pub fn push_null(&mut self)

Add a Null value in array

pub fn push_bool(&mut self, value: Option<bool>)

Add a boolean value in array, if it’s not a boolean array failled

pub fn push_i32(&mut self, value: Option<i32>)

Add a i32 value in array, if it’s not a integer array failled

pub fn push_f32(&mut self, value: Option<f32>)

Add a f32 value in array, if it’s not a float array failled

pub fn push_string(&mut self, value: String)

Add a string value in array, if it’s not a string array failled

pub fn push_vecbool(&mut self, value: Vec<Option<bool>>) -> Result<()>

Add a vector of bool value in array, if it’s not a vector of bool array failled

pub fn push_veci32(&mut self, value: Vec<Option<i32>>) -> Result<()>

Add a vector of integer value in array, if it’s not a vector of integer array failled

pub fn push_vecf32(&mut self, value: Vec<Option<f32>>) -> Result<()>

Add a vector of float value in array, if it’s not a vector of float array failled

pub fn push_vecstring(&mut self, value: Vec<Option<String>>) -> Result<()>

Add a vector of string value in array, if it’s not a vector of string array failled

pub fn into_arc(self) -> Arc<dyn Array>

Convert ColumnData in Arrow2 array